moved class events to slack_user_controller.rb
